requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
matplotlib>=3.4.2
Pillow>=8.3.1
tqdm>=4.62.0